+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ## Added
 
+## [0.5.1] - 2023-11-19
+## Added
+- Added `Count(*)` by parsing it as `Count(1)`
+
+### Changed
+- Add a warning when the textbook gaussian mechanism is not really applied [#190](https://github.com/Qrlew/qrlew/issues/190)
+
 ## [0.5.0] - 2023-11-19
 ### Changed
 - Breaking name change: Protected Entity -> Privacy Unit + PEP -> PUP [#189](https://github.com/Qrlew/qrlew/issues/189)